--- a/lua/theme.lua
+++ b/lua/theme.lua
@@ -4,7 +4,7 @@ vim.g.nvchad_theme = chad_theme
 local present, base16 = pcall(require, "base16")
 
 if present then
-    base16(base16.themes[chad_theme], true)
+    base16(base16.themes(chad_theme), true)
     require "highlights"
     return true
 else