Improvements in startup time

Before:
```
require("theme"): 5.38458ms (24.332083 -> 29.716666)
    require("user_config"): 0.030104ms (24.402135 -> 24.432239)
    require("base16"): 2.62813ms (24.442708 -> 27.070833)
      require("nvim"): 0.317813ms (26.695260 -> 27.013073)
    require("highlights"): 0.663593ms (29.051198 -> 29.714791)
      require("themes/onedark"): 0.056563ms (29.265364 -> 29.321927)
```

After:
```
require("theme"): 3.69021ms (23.268958 -> 26.959167)
    require("user_config"): 0.026771ms (23.325677 -> 23.352448)
    require("base16"): 1.0087ms (23.362031 -> 24.370729)
      require("nvim"): 0.305208ms (24.063646 -> 24.368854)
    require("themes/onedark-base16"): 0.058646ms (24.371979 -> 24.430625)
    require("highlights"): 0.627188ms (26.330312 -> 26.957500)
      require("themes/onedark"): 0.057448ms (26.521250 -> 26.578698)
```

Benchmark taken with profiler.nvim

This should be only merged after https://github.com/siduck76/nvim-base16.lua/pull/3 is merged